Stage 01: correct, secure and honest map-painter
Completes every task in stages/STAGE-01-prototype.md.

Secure the journey endpoints
  [Authorize] on JourneyController; the player is resolved from the JWT via
  IUserContextHelper instead of a client-supplied PlayerId, which previously let
  anyone read or mutate any player's map without a token.

Swept-path reveal
  A walk between two syncs used to lose every cell in between. SyncRequest now takes
  an array of positions and FogService sweeps a supercover line across the grid, so
  sampling rate no longer changes the map. RevealRadius 0 -> 1, one batch insert
  instead of per-cell, MaxNewCellsPerSync 9 -> 2000 so a real walk is not truncated.

  The supercover takes *both* cells at an exact corner crossing. Taking one made the
  revealed set depend on which way you walked the street; a 13x13 endpoint sweep now
  covers that.

Anti-cheat
  Swept reveal amplifies both exploits it guards against - a phone drifting on a desk
  would paint a neighbourhood and a car would paint a city. TraceValidator gates the
  sweep on accuracy (25 m), displacement ratio (0.3), dwell (20 m / 180 s) and speed
  (5 m/s). Speed is graded on net displacement, not path length, so GPS jitter does
  not flag a slow walk as a vehicle. Dwell also needed a cross-sync check, since a
  phone syncing every 30 s never accumulates 180 s of evidence within one batch.

Background sync
  Sends the whole batch rather than only the last fix. Token refresh is extracted to
  helpers/tokenRefresh.ts, free of React and expo-router - the interceptor version
  calls router.replace, which throws in a background task. Failed batches persist to
  AsyncStorage and retry, spaced so the server's own 2 s cooldown does not reject the
  backlog. A long walk no longer 401s into an empty catch and silently loses ground.

XP curve
  RuneScape curve, table to level 200 with the formula beyond, uncapped. Player.Xp now
  means cumulative lifetime XP with Level derived from it, replacing a within-level
  remainder that could not survive a curve change. Widened int -> long, which an
  uncapped curve needs (int caps out near level 126); the HUD's level*100 bar was
  pegging at 100% under cumulative XP and now mirrors the server curve.

Housekeeping
  Both NU1903 packages were transitive, so they are pinned directly (Microsoft.OpenApi
  3.10.2, SSH.NET 2026.0.0); Swashbuckle already wanted >= 3.10.2, so the 2.4.1 was a
  stale floor. Debug seed guard checked for financemanagercontainer, a template
  leftover that never matched. Seq endpoint reads SEQ_URL. Stale StreetImportService
  docstring and three "street progress" permission strings removed - streets are gone.

Also adds AppDbContextFactory so dotnet ef works without Docker; a Debug build starts
a Testcontainer at host startup, which EF's default path would otherwise require.

Verification: build clean (0 errors, 0 warnings), 64 unit tests pass, criteria 1-9
verified. Criterion 10 (app builds and renders) is NOT verified - geoslayer.app has no
node_modules here, so the app was never typechecked. The Testcontainers integration
tests did not run either; Docker is unavailable. Both are queued in stages/STATUS.md.

+++ GeoSlayer.Domain/DTOs/Auth/Responses/LoginUserResponse.cs
@@ -7,6 +7,6 @@
         public int PlayerId { get; set; }
         public string Username { get; set; } = "";
         public int Level { get; set; }
-        public int Xp { get; set; }
+        public long Xp { get; set; }
     }
 }
